Moved styling for interactive UI on recipe feed to dedicated CSS module
