+++ src/components/Feed/PostInteraction/PostInteraction.module.css
@@ -32,6 +32,40 @@
     align-items: center;
 }
 
+/*Recipe Image Container -------------------------------------------------*/
+.recipe--imgContainer {
+  background-color: #edf4e4;
+  background-image: var(--bg-image);
+  /*Dynamically set recipe image to background image for control over dimensions*/
+  width: 100%;
+  padding-top: 100%;
+  /*Make image square in mobile view*/
+  background-repeat: no-repeat;
+  background-position: center;
+  background-size: cover;
+  background-blend-mode: multiply;
+  border-radius: 25px;
+  box-shadow: 8px 6px 4px rgba(54, 68, 68, 0.25);
+  transition: all .2s ease-in-out;
+}
+
+.recipe--imgContainer:hover {
+  cursor: pointer;
+}
+
+@media (min-width: 520px) {
+  .recipe--imgContainer {
+    width: 15rem;
+    height: 15rem;
+    padding: 0;
+  }
+
+  .recipe--imgContainer:hover {
+    transform: scale(1.02);
+  }
+}
+/*------------------------------------------------------------------------*/
+
 /*Favorite Icon------------------------------------------*/
 .favstar {
     /* Modify size here: */
